requirements_dev.txt
# torch
accelerate
jupyter
kaleido
matplotlib
numpy
pandas
plotly
python-dotenv
seaborn
sentence-transformers
transformers
tiktoken
blobfile
sentencepiece
protobuf

adjustText
gradio
aiolimiter
openai
anthropic
google-genai